Partly move getEdges into EdgeTableWrapper
